rt: move `enter` into `context`
