+++ tokio/src/task/local.rs
@@ -890,7 +890,7 @@ impl<T: Future> Future for RunUntil<'_, T> {
                 .waker
                 .register_by_ref(cx.waker());
 
-            let _no_blocking = crate::runtime::enter::disallow_block_in_place();
+            let _no_blocking = crate::runtime::context::disallow_block_in_place();
             let f = me.future;
 
             if let Poll::Ready(output) = f.poll(cx) {
